Describe ParseReference's dep on GetStore
Calls to ParseReference() eventually end up calling GetStore(), and
GetStore() needs more information than a reference can tell it, so a few
test cases that appear to be valid at first glance still fail to parse.

Add some more detail on why that happens: in order to match a valid
store, that store needs to have been previously opened by the calling
application, but since the store is in a temporary directory created
just for the test, and we aren't opening that location with GetStore()
first, it's expected to fail.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>