Make creator ID and timestamp explicit fields of untrustedSignature
Instead of silently embedding values in untrustedSignature.MarshalJSON
(and having to add a marshalJSONWithvariables to work around this),
make the creator ID and timestamp explicit fields of untrustedSignature,
and MarshalJSON a simple marshaller of existing data.

The values are now filled by calling newUntrustedSignature.

Now that the fields are explicit, we can also record them by
untrustedSignature.UnmarshalJSON.

This also explicitly defines the timestamp to be an integer, instead of
allowing floating-point values, because the JSON float64 is not precise
enough for nanosecond timstamps. For now, we reject fractional values,
which will allow us to record the nanosecond part separately in the
future if it became necessary.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>