docker: daemon: decompress blobs transparently in GetBlob
This is necessary because inside the docker-daemon (soon to be Docker
archive) transport, the provided DigestInfo's digest is *not* the digest
of the blob but is rather the DiffID of the layer. This results in
issues when converting a docker-daemon image to an OCI image and then
back again -- the compression copying code doesn't know what to do
because the digest of the blob and the blob's "digest" in DigestInfo
don't match.

Fix it by always silently decompressing blobs that come from Docker
archive transports.

Signed-off-by: Aleksa Sarai <asarai@suse.de>