Fail in HasBlob on unexpected HTTP statuses
The code originates in PutBlob, where unexpected HTTP statuses made us
abort and not even try to upload; so, it seems that HasBlob should
similarly return an error (possibly causing the caller to abort)
similarly.  The caller can of course still decide to ignore unexpected
errors, but this is not an “all is well, blob has not been uploaded”
situation (or at least we have been happy enough not to treat it as such a
situation in PutBlob so far).

Signed-off-by: Miloslav Trmač <mitr@redhat.com>