Turn dockerClient.ping into dockerClient.detectProperties()
This only moves the client.scheme == "" check inside the method; but the
naming will make it more appropriate to call the method from
dockerImagesource / dockerImageDestination when they need to depend on
the results of the properties detection in methods which may be called
before dockerClient.makeRequest is ever called (detecting the properties
as a side effect).

Does not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>